+++ zone/CMakeLists.txt
@@ -25,7 +25,6 @@ SET(zone_sources
     combat_record.cpp
     command.cpp
     corpse.cpp
-    data_bucket.cpp
     doors.cpp
     dialogue_window.cpp
     dynamic_zone.cpp
@@ -195,7 +194,6 @@ SET(zone_headers
     command.h
     common.h
     corpse.h
-    data_bucket.h
     doors.h
     dialogue_window.h
     dynamic_zone.h
