+++ zone/client.cpp
@@ -39,7 +39,7 @@ extern volatile bool RunLoops;
 #include "../common/strings.h"
 #include "../common/data_verification.h"
 #include "../common/profanity_manager.h"
-#include "data_bucket.h"
+#include "../common/data_bucket.h"
 #include "dynamic_zone.h"
 #include "expedition_request.h"
 #include "position.h"
